functions/process_upload/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
pandas
